+++ src/library/Parser/string/StringParser.ts
@@ -1,5 +1,5 @@
-import type Parser from '..';
 import type ParsingError from '../ParsingError';
+import type Parser from '../definition';
 
 type StringParser<
   SomeParsedOutput,
